requirements.txt
telethon==1.26.1
pandas==1.5.3
openpyxl==3.0.10
tqdm==4.64.1
networkx==3.0
matplotlib==3.6.3
python-louvain==0.16
streamlit==1.23.1
streamlit-extras==0.2.7
pycryptodome==3.17
pywebview==4.2.2
psutil==5.9.5
nuitka==1.7.3